basic keyboard and setting is created
DESCRIPTION:
feachers:
- check memeber ship
- loger
- auto add handler
