--- .VSCodeCounter/2024-08-16_13-21-09/details.md
@@ -1,39 +1,0 @@
-# Details
-
-Date : 2024-08-16 13:21:09
-
-Directory /home/saman/gits/PRJ/vscode_github
-
-Total : 24 files,  2146 codes, 337 comments, 234 blanks, all 2717 lines
-
-[Summary](results.md) / Details / [Diff Summary](diff.md) / [Diff Details](diff-details.md)
-
-## Files
-| filename | language | code | comment | blank | total |
-| :--- | :--- | ---: | ---: | ---: | ---: |
-| [Dockerfile](/Dockerfile) | Docker | 9 | 7 | 6 | 22 |
-| [Markups.py](/Markups.py) | Python | 29 | 4 | 2 | 35 |
-| [Pipfile](/Pipfile) | toml | 11 | 0 | 4 | 15 |
-| [Pipfile.lock](/Pipfile.lock) | JSON | 206 | 0 | 1 | 207 |
-| [configs/auth.py](/configs/auth.py) | Python | 12 | 3 | 1 | 16 |
-| [configs/config.py](/configs/config.py) | Python | 34 | 14 | 2 | 50 |
-| [database/db_creation.py](/database/db_creation.py) | Python | 149 | 11 | 5 | 165 |
-| [database/db_functions.py](/database/db_functions.py) | Python | 70 | 4 | 13 | 87 |
-| [database/db_reserve.py](/database/db_reserve.py) | Python | 111 | 23 | 2 | 136 |
-| [database/db_timing.py](/database/db_timing.py) | Python | 68 | 6 | 0 | 74 |
-| [database/db_transactions.py](/database/db_transactions.py) | Python | 76 | 8 | 4 | 88 |
-| [database/db_users.py](/database/db_users.py) | Python | 173 | 18 | 4 | 195 |
-| [functions/custom_functions.py](/functions/custom_functions.py) | Python | 273 | 50 | 61 | 384 |
-| [logrotate.conf](/logrotate.conf) | Properties | 12 | 1 | 1 | 14 |
-| [main.py](/main.py) | Python | 740 | 127 | 88 | 955 |
-| [message_and_text/bot_message_functions.py](/message_and_text/bot_message_functions.py) | Python | 21 | 2 | 4 | 27 |
-| [message_and_text/bot_messages.py](/message_and_text/bot_messages.py) | Python | 17 | 6 | 1 | 24 |
-| [message_and_text/text.py](/message_and_text/text.py) | Python | 52 | 5 | 13 | 70 |
-| [readme.md](/readme.md) | Markdown | 9 | 0 | 0 | 9 |
-| [requirements.txt](/requirements.txt) | pip requirements | 5 | 0 | 0 | 5 |
-| [states.py](/states.py) | Python | 18 | 0 | 4 | 22 |
-| [tests/test.py](/tests/test.py) | Python | 29 | 44 | 13 | 86 |
-| [tests/test2.py](/tests/test2.py) | Python | 15 | 4 | 2 | 21 |
-| [tests/test3.py](/tests/test3.py) | Python | 7 | 0 | 3 | 10 |
-
-[Summary](results.md) / Details / [Diff Summary](diff.md) / [Diff Details](diff-details.md)
