--- .VSCodeCounter/2024-08-16_13-21-09/diff-details.md
@@ -1,15 +1,0 @@
-# Diff Details
-
-Date : 2024-08-16 13:21:09
-
-Directory /home/saman/gits/PRJ/vscode_github
-
-Total : 0 files,  0 codes, 0 comments, 0 blanks, all 0 lines
-
-[Summary](results.md) / [Details](details.md) / [Diff Summary](diff.md) / Diff Details
-
-## Files
-| filename | language | code | comment | blank | total |
-| :--- | :--- | ---: | ---: | ---: | ---: |
-
-[Summary](results.md) / [Details](details.md) / [Diff Summary](diff.md) / Diff Details
