--- .VSCodeCounter/2024-08-16_13-21-09/diff.csv
@@ -1,2 +1,0 @@
-"filename", "language", "", "comment", "blank", "total"
-"Total", "-", , 0, 0, 0
