--- .VSCodeCounter/2024-08-16_13-21-09/diff.txt
@@ -1,22 +1,0 @@
-Date : 2024-08-16 13:21:09
-Directory : /home/saman/gits/PRJ/vscode_github
-Total : 0 files,  0 codes, 0 comments, 0 blanks, all 0 lines
-
-Languages
-+----------+------------+------------+------------+------------+------------+
-| language | files      | code       | comment    | blank      | total      |
-+----------+------------+------------+------------+------------+------------+
-+----------+------------+------------+------------+------------+------------+
-
-Directories
-+------+------------+------------+------------+------------+------------+
-| path | files      | code       | comment    | blank      | total      |
-+------+------------+------------+------------+------------+------------+
-+------+------------+------------+------------+------------+------------+
-
-Files
-+----------+----------+------------+------------+------------+------------+
-| filename | language | code       | comment    | blank      | total      |
-+----------+----------+------------+------------+------------+------------+
-| Total    |          |          0 |          0 |          0 |          0 |
-+----------+----------+------------+------------+------------+------------+
