--- .VSCodeCounter/2024-08-16_13-21-09/results.json
@@ -1,1 +1,0 @@
-{"file:///home/saman/gits/PRJ/vscode_github/tests/test.py":{"language":"Python","code":29,"comment":44,"blank":13},"file:///home/saman/gits/PRJ/vscode_github/tests/test3.py":{"language":"Python","code":7,"comment":0,"blank":3},"file:///home/saman/gits/PRJ/vscode_github/tests/test2.py":{"language":"Python","code":15,"comment":4,"blank":2},"file:///home/saman/gits/PRJ/vscode_github/requirements.txt":{"language":"pip requirements","code":5,"comment":0,"blank":0},"file:///home/saman/gits/PRJ/vscode_github/message_and_text/bot_message_functions.py":{"language":"Python","code":21,"comment":2,"blank":4},"file:///home/saman/gits/PRJ/vscode_github/message_and_text/text.py":{"language":"Python","code":52,"comment":5,"blank":13},"file:///home/saman/gits/PRJ/vscode_github/Pipfile":{"language":"toml","code":11,"comment":0,"blank":4},"file:///home/saman/gits/PRJ/vscode_github/main.py":{"language":"Python","code":740,"comment":127,"blank":88},"file:///home/saman/gits/PRJ/vscode_github/readme.md":{"language":"Markdown","code":9,"comment":0,"blank":0},"file:///home/saman/gits/PRJ/vscode_github/configs/config.py":{"language":"Python","code":34,"comment":14,"blank":2},"file:///home/saman/gits/PRJ/vscode_github/message_and_text/bot_messages.py":{"language":"Python","code":17,"comment":6,"blank":1},"file:///home/saman/gits/PRJ/vscode_github/logrotate.conf":{"language":"Properties","code":12,"comment":1,"blank":1},"file:///home/saman/gits/PRJ/vscode_github/Markups.py":{"language":"Python","code":29,"comment":4,"blank":2},"file:///home/saman/gits/PRJ/vscode_github/database/db_users.py":{"language":"Python","code":173,"comment":18,"blank":4},"file:///home/saman/gits/PRJ/vscode_github/configs/auth.py":{"language":"Python","code":12,"comment":3,"blank":1},"file:///home/saman/gits/PRJ/vscode_github/database/db_timing.py":{"language":"Python","code":68,"comment":6,"blank":0},"file:///home/saman/gits/PRJ/vscode_github/database/db_reserve.py":{"language":"Python","code":111,"comment":23,"blank":2},"file:///home/saman/gits/PRJ/vscode_github/database/db_creation.py":{"language":"Python","code":149,"comment":11,"blank":5},"file:///home/saman/gits/PRJ/vscode_github/functions/custom_functions.py":{"language":"Python","code":273,"comment":50,"blank":61},"file:///home/saman/gits/PRJ/vscode_github/database/db_transactions.py":{"language":"Python","code":76,"comment":8,"blank":4},"file:///home/saman/gits/PRJ/vscode_github/database/db_functions.py":{"language":"Python","code":70,"comment":4,"blank":13},"file:///home/saman/gits/PRJ/vscode_github/Pipfile.lock":{"language":"JSON","code":206,"comment":0,"blank":1},"file:///home/saman/gits/PRJ/vscode_github/Dockerfile":{"language":"Docker","code":9,"comment":7,"blank":6},"file:///home/saman/gits/PRJ/vscode_github/states.py":{"language":"Python","code":18,"comment":0,"blank":4}}
