--- .VSCodeCounter/2024-08-16_13-21-09/results.md
@@ -1,33 +1,0 @@
-# Summary
-
-Date : 2024-08-16 13:21:09
-
-Directory /home/saman/gits/PRJ/vscode_github
-
-Total : 24 files,  2146 codes, 337 comments, 234 blanks, all 2717 lines
-
-Summary / [Details](details.md) / [Diff Summary](diff.md) / [Diff Details](diff-details.md)
-
-## Languages
-| language | files | code | comment | blank | total |
-| :--- | ---: | ---: | ---: | ---: | ---: |
-| Python | 18 | 1,894 | 329 | 222 | 2,445 |
-| JSON | 1 | 206 | 0 | 1 | 207 |
-| Properties | 1 | 12 | 1 | 1 | 14 |
-| toml | 1 | 11 | 0 | 4 | 15 |
-| Markdown | 1 | 9 | 0 | 0 | 9 |
-| Docker | 1 | 9 | 7 | 6 | 22 |
-| pip requirements | 1 | 5 | 0 | 0 | 5 |
-
-## Directories
-| path | files | code | comment | blank | total |
-| :--- | ---: | ---: | ---: | ---: | ---: |
-| . | 24 | 2,146 | 337 | 234 | 2,717 |
-| . (Files) | 9 | 1,039 | 139 | 106 | 1,284 |
-| configs | 2 | 46 | 17 | 3 | 66 |
-| database | 6 | 647 | 70 | 28 | 745 |
-| functions | 1 | 273 | 50 | 61 | 384 |
-| message_and_text | 3 | 90 | 13 | 18 | 121 |
-| tests | 3 | 51 | 48 | 18 | 117 |
-
-Summary / [Details](details.md) / [Diff Summary](diff.md) / [Diff Details](diff-details.md)
