--- .VSCodeCounter/2024-08-16_13-21-09/results.txt
@@ -1,60 +1,0 @@
-Date : 2024-08-16 13:21:09
-Directory : /home/saman/gits/PRJ/vscode_github
-Total : 24 files,  2146 codes, 337 comments, 234 blanks, all 2717 lines
-
-Languages
-+------------------+------------+------------+------------+------------+------------+
-| language         | files      | code       | comment    | blank      | total      |
-+------------------+------------+------------+------------+------------+------------+
-| Python           |         18 |      1,894 |        329 |        222 |      2,445 |
-| JSON             |          1 |        206 |          0 |          1 |        207 |
-| Properties       |          1 |         12 |          1 |          1 |         14 |
-| toml             |          1 |         11 |          0 |          4 |         15 |
-| Markdown         |          1 |          9 |          0 |          0 |          9 |
-| Docker           |          1 |          9 |          7 |          6 |         22 |
-| pip requirements |          1 |          5 |          0 |          0 |          5 |
-+------------------+------------+------------+------------+------------+------------+
-
-Directories
-+------------------------------------------------------------------------------+------------+------------+------------+------------+------------+
-| path                                                                         | files      | code       | comment    | blank      | total      |
-+------------------------------------------------------------------------------+------------+------------+------------+------------+------------+
-| .                                                                            |         24 |      2,146 |        337 |        234 |      2,717 |
-| . (Files)                                                                    |          9 |      1,039 |        139 |        106 |      1,284 |
-| configs                                                                      |          2 |         46 |         17 |          3 |         66 |
-| database                                                                     |          6 |        647 |         70 |         28 |        745 |
-| functions                                                                    |          1 |        273 |         50 |         61 |        384 |
-| message_and_text                                                             |          3 |         90 |         13 |         18 |        121 |
-| tests                                                                        |          3 |         51 |         48 |         18 |        117 |
-+------------------------------------------------------------------------------+------------+------------+------------+------------+------------+
-
-Files
-+------------------------------------------------------------------------------+------------------+------------+------------+------------+------------+
-| filename                                                                     | language         | code       | comment    | blank      | total      |
-+------------------------------------------------------------------------------+------------------+------------+------------+------------+------------+
-| /home/saman/gits/PRJ/vscode_github/Dockerfile                                | Docker           |          9 |          7 |          6 |         22 |
-| /home/saman/gits/PRJ/vscode_github/Markups.py                                | Python           |         29 |          4 |          2 |         35 |
-| /home/saman/gits/PRJ/vscode_github/Pipfile                                   | toml             |         11 |          0 |          4 |         15 |
-| /home/saman/gits/PRJ/vscode_github/Pipfile.lock                              | JSON             |        206 |          0 |          1 |        207 |
-| /home/saman/gits/PRJ/vscode_github/configs/auth.py                           | Python           |         12 |          3 |          1 |         16 |
-| /home/saman/gits/PRJ/vscode_github/configs/config.py                         | Python           |         34 |         14 |          2 |         50 |
-| /home/saman/gits/PRJ/vscode_github/database/db_creation.py                   | Python           |        149 |         11 |          5 |        165 |
-| /home/saman/gits/PRJ/vscode_github/database/db_functions.py                  | Python           |         70 |          4 |         13 |         87 |
-| /home/saman/gits/PRJ/vscode_github/database/db_reserve.py                    | Python           |        111 |         23 |          2 |        136 |
-| /home/saman/gits/PRJ/vscode_github/database/db_timing.py                     | Python           |         68 |          6 |          0 |         74 |
-| /home/saman/gits/PRJ/vscode_github/database/db_transactions.py               | Python           |         76 |          8 |          4 |         88 |
-| /home/saman/gits/PRJ/vscode_github/database/db_users.py                      | Python           |        173 |         18 |          4 |        195 |
-| /home/saman/gits/PRJ/vscode_github/functions/custom_functions.py             | Python           |        273 |         50 |         61 |        384 |
-| /home/saman/gits/PRJ/vscode_github/logrotate.conf                            | Properties       |         12 |          1 |          1 |         14 |
-| /home/saman/gits/PRJ/vscode_github/main.py                                   | Python           |        740 |        127 |         88 |        955 |
-| /home/saman/gits/PRJ/vscode_github/message_and_text/bot_message_functions.py | Python           |         21 |          2 |          4 |         27 |
-| /home/saman/gits/PRJ/vscode_github/message_and_text/bot_messages.py          | Python           |         17 |          6 |          1 |         24 |
-| /home/saman/gits/PRJ/vscode_github/message_and_text/text.py                  | Python           |         52 |          5 |         13 |         70 |
-| /home/saman/gits/PRJ/vscode_github/readme.md                                 | Markdown         |          9 |          0 |          0 |          9 |
-| /home/saman/gits/PRJ/vscode_github/requirements.txt                          | pip requirements |          5 |          0 |          0 |          5 |
-| /home/saman/gits/PRJ/vscode_github/states.py                                 | Python           |         18 |          0 |          4 |         22 |
-| /home/saman/gits/PRJ/vscode_github/tests/test.py                             | Python           |         29 |         44 |         13 |         86 |
-| /home/saman/gits/PRJ/vscode_github/tests/test2.py                            | Python           |         15 |          4 |          2 |         21 |
-| /home/saman/gits/PRJ/vscode_github/tests/test3.py                            | Python           |          7 |          0 |          3 |         10 |
-| Total                                                                        |                  |      2,146 |        337 |        234 |      2,717 |
-+------------------------------------------------------------------------------+------------------+------------+------------+------------+------------+
