--- .vscode/settings.json
@@ -1,28 +1,0 @@
-{
-    "cSpell.words": [
-        "CURDATE",
-        "dotenv",
-        "jalali",
-        "joinchat",
-        "linkgp",
-        "lιnĸ",
-        "naмe",
-        "patern",
-        "saaman",
-        "Shamsi",
-        "telebot",
-        "мeмвer",
-        "آپدیت",
-        "ادمین",
-        "ازای",
-        "اکانت",
-        "بنرها",
-        "رزور",
-        "کارکتر",
-        "کنین",
-        "میگرد",
-        "هزارتومان",
-        "هندلر"
-    ],
-    "cSpell.language": "en,fa-IR"
-}
