+++ app/messages/fa.py
@@ -1,0 +1,21 @@
+MESSAGES = {
+    # General
+    "start.welcome": "سلام 👋 به ربات خوش اومدی!",
+    "admin.panel": "📊 خوش اومدی به پنل ادمین.",
+    "error.not_member": "❗️برای استفاده از ربات باید عضو کانال {channel} باشی.",
+
+    # User buttons
+    "btn.free_times": "🕒 ساعت‌های خالی",
+    "btn.my_reservations": "📆 مشاهده رزروها",
+    "btn.convert_points": "🎁 تبدیل امتیاز",
+    "btn.profile": "👤 حساب کاربری",
+    "btn.banner": "🖼 ساخت بنر",
+    "btn.support": "📞 پشتیبانی",
+
+    # Admin Buttons 
+    "btn.admin.bot_setting": "⚙️ تنظیمات ربات",
+    "btn.admin.free_time": "📆 مدیریت رزروها",
+    "btn.admin.user_list": "👥 لیست کاربران",
+    "btn.admin.income": "💰 درآمد",
+    # "btn.admin.send_msg_to_all":"",
+}
