+++ old_bot/configs/basic_info.py
