+++ old_bot/functions/format_patern.py
