+++ old_bot/message_and_text/bot_message_functions.py
