--- readme.md
@@ -1,16 +1,0 @@
-this is a telegram robot working with 'TeleBot' and python .
-for this code to work you need this :
-for docker we use : 'python:3.9.19-slim'
-```python
-pip install schedule --break-system-packages
-pip install telebot --break-system-packages
-pip install mysql-connector-python --break-system-packages
-pip install convertdate --break-system-packages
-```
-just testing git actions
-
-hi
-test yml
-make env file
-
-test
